+++ omechu-app/src/app/mypage/api/profile.ts
@@ -1,6 +1,7 @@
 import axiosInstance from "@/lib/api/axios";
-import { ProfileType } from "../types/profileType";
+
 import { useAuthStore } from "@/lib/stores/auth.store";
+import { ProfileType } from "../types/profileType";
 
 export class ProfileApiError extends Error {
   constructor(
