+++ omechu-app/src/app/mypage/user-info-edit/[step]/EditStart.tsx
@@ -25,9 +25,9 @@ export default function EditStart() {
         }
       />{" "}
       <main className="flex min-h-[calc(100vh-3.1rem)] w-full flex-col items-center overflow-y-scroll overscroll-none scroll-smooth px-4">
-        <section className="mb-28 mt-36 flex flex-col gap-5 text-center">
+        <section className="mt-36 mb-28 flex flex-col gap-5 text-center">
           <div className="text-2xl font-medium">기본 정보 입력하기</div>
-          <span className="whitespace-pre font-normal text-grey-normal-active">
+          <span className="text-grey-normal-active font-normal whitespace-pre">
             기본정보를 저장하여 {"\n"} 더 정교한 메뉴 추천을 받아보세요!
           </span>
           {/* prettier formatOnSave 때문에 whitespace-pre 미적용 -> 추후 수정 */}
@@ -37,7 +37,7 @@ export default function EditStart() {
             onClick={() =>
               router.push(`/mypage/user-info-edit/${indexToSlug[1]}`)
             }
-            className="flex h-16 w-48 items-center justify-center rounded-md bg-primary-normal p-5 text-2xl font-medium text-white hover:bg-primary-normal-hover active:bg-primary-normal-active"
+            className="bg-primary-normal hover:bg-primary-normal-hover active:bg-primary-normal-active flex h-16 w-48 items-center justify-center rounded-md p-5 text-2xl font-medium text-white"
           >
             시작하기
           </button>
